Add Devise Case Study
This bit of documentation shows exactly how powerful `rails-decorators`
can be for people who are somewhat new to Rails and/or decorating
classes. The examples given in the docs are great because they are
somewhat generic, but I felt like a more concrete example would be
useful as well so people browsing the repo can clearly see the benefits
of using this gem, even in non-Workarea applications.